requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
uvicorn[standard]==0.38.0

httpx==0.28.1
sentence-transformers==5.1.2
scikit-learn==1.6.1
rapidfuzz==3.13.0

python-dotenv==1.2.1
pydantic-settings==2.11.0

openai==1.72.0
faiss-cpu==1.13.0
